refactor: move archival code to own directory
