+++ baker/archival/buildLocalArchivalBake.ts
@@ -8,16 +8,16 @@ import yargs from "yargs"
 import { hideBin } from "yargs/helpers"
 import fs from "fs-extra"
 import path, { normalize } from "path"
-import * as db from "../db/db.js"
-import { bakeSingleGrapherPageForArchival } from "./GrapherBaker.js"
+import * as db from "../../db/db.js"
+import { bakeSingleGrapherPageForArchival } from "../GrapherBaker.js"
 import { keyBy } from "lodash"
-import { getAllImages } from "../db/model/Image.js"
-import { getChartConfigBySlug } from "../db/model/Chart.js"
-import findProjectBaseDir from "../settings/findBaseDir.js"
+import { getAllImages } from "../../db/model/Image.js"
+import { getChartConfigBySlug } from "../../db/model/Chart.js"
+import findProjectBaseDir from "../../settings/findBaseDir.js"
 import { AssetMap } from "@ourworldindata/types"
-import { getVariableData } from "../db/model/Variable.js"
+import { getVariableData } from "../../db/model/Variable.js"
 import dayjs from "dayjs"
-import { hashBase36, hashBase36FromStream } from "../serverUtils/hash.js"
+import { hashBase36, hashBase36FromStream } from "../../serverUtils/hash.js"
 import * as Sentry from "@sentry/node"
 
 const DATE_TIME_FORMAT = "YYYYMMDD-HHmmss"
